% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_regression.R
\name{adaptive_regression}
\alias{adaptive_regression}
\title{Adaptive Regression}
\usage{
adaptive_regression(df, t, y, B, g)
}
\arguments{
\item{df}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame}

\item{t}{Name of the column of type Date containing the dates}

\item{y}{Name of the column of type Numeric containing counts or percentages}

\item{B}{Baseline parameter. The baseline length is the number of days to
which each liner model is fit.}

\item{g}{Guardband parameter. The guardband length is the number of days
separating the baseline from the current date in consideration for alerting.}
}
\value{
A data frame.
}
\description{
Adaptive Regression helper function for Adaptive Multiple Regression.
}
\keyword{internal}
