% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rmd_template_gui.R
\name{add_rmd_template_gui}
\alias{add_rmd_template_gui}
\title{Add Rnssp template}
\usage{
add_rmd_template_gui()
}
\description{
Wrapper around \code{\link[=add_rmd_template]{add_rmd_template()}} to add Rnssp templates from
the Rnssp templates Github repository.
}
\keyword{internal}
