% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_nbinom.R
\name{alert_nbinom}
\alias{alert_nbinom}
\title{Negative binomial detection algorithm for weekly counts}
\usage{
alert_nbinom(df, t = date, y = count, baseline_end)
}
\arguments{
\item{df}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame.}

\item{t}{Name of the column of type Date containing the dates}

\item{y}{Name of the column of type Numeric containing counts}

\item{baseline_end}{Object of type Date defining the end of the
baseline/training period}
}
\value{
A data frame with model estimates, lower and upper prediction
interval bounds, and a binary alarm indicator field
}
\description{
The negative binomial regression algorithm fits a negative binomial
regression model with a time term and order 1 Fourier terms to a baseline
period that spans 2 or more years. Inclusion of Fourier terms in the model
is intended to account for seasonality common in multi-year weekly time
series of counts. Order 1 sine and cosine terms are included to account for
annual seasonality that is common to syndromes and diseases such as influenza,
RSV, and norovirus. Each baseline model is used to make weekly forecasts
for all weeks following the baseline period. Upper and lower 95% prediction
interval bounds are computed for each week in the prediction period.
Alarms are signaled for any week during for which
weekly counts fall above the upper bound of the prediction interval.
}
