% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{askme}
\alias{askme}
\title{Password Prompt Utility}
\usage{
askme(prompt = "Please enter your password: ")
}
\value{
a character string
}
\description{
Prompt the user for a password. This function is a wrapper for the \code{\link[getPass]{getPass}} function.
}
\examples{
password <- askme()
}
\seealso{
\code{\link[getPass]{getPass}}
}
