% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{change_dates}
\alias{change_dates}
\title{Utility to Change the Start and/Or End Dates in an ESSENCE API URL}
\usage{
change_dates(url, start_date = NULL, end_date = NULL)
}
\arguments{
\item{url}{a character string of ESSENCE API URL}

\item{start_date}{a date object or a character string in date format (e.g. "2019-08-01")}

\item{end_date}{a date object or a character string in date format (e.g. "2020-08-01")}
}
\value{
a character string
}
\description{
Change the start and/or end dates in an ESSENCE API URL
}
\examples{
library(Rnssp)

url <- "https://essence.syndromicsurveillance.org/nssp_essence/api/alerts/hospitalSyndromeAlerts?
end_date=31Jan2021&start_date=29Jan2021"

url <- url \%>\% gsub("\n", "", .)

# Change start date to January 15, 2021
url \%>\% change_dates(start_date = "2021-01-15")

# Change end date to February 15, 2021
url \%>\% change_dates(end_date = "2021-02-15")

# Change start date to January 15, 2021 and end date to February 15, 2021
url \%>\% change_dates(start_date = "2021-01-15", end_date = "2021-02-15")

# Change end and start dates to respectively "current" and "current - 7 days"
url \%>\% change_dates(start_date = Sys.Date() - 7, end_date = Sys.Date())
}
