% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Removes stopwords and common patterns of Chief Complaints}
\usage{
clean_text(tbl)
}
\arguments{
\item{tbl}{the input dataframe}
}
\value{
A transformed tibble dataframe with clean CC and CCDD text fields
}
\description{
Converts a dataframe or tibble into a data.table object,
removes stop words and patterns identified by a series of regular expressions.
}
\examples{
\dontrun{
data <- api_data \%>\%
  clean_text()
}
}
