% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_sf.R
\docType{data}
\name{county_sf}
\alias{county_sf}
\title{US Counties Shapefile}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 3142 rows and 10 columns.
}
\usage{
county_sf
}
\description{
Useful \code{sf} object shapefile for US Counties.
}
\details{
\preformatted{
Rows: 3,142
Columns: 10
$ STATEFP  <fct> 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, 01, ...
$ COUNTYFP <fct> 001, 003, 005, 007, 009, 011, 013, 015, 017, 019, 021, 023, 025, 027, 029, 031, 033, 035, 037, 039, 041, 043, 045, 047, 049, 051, 05...
$ COUNTYNS <fct> 00161526, 00161527, 00161528, 00161529, 00161530, 00161531, 00161532, 00161533, 00161534, 00161535, 00161536, 00161537, 00161538, 00...
$ AFFGEOID <fct> 0500000US01001, 0500000US01003, 0500000US01005, 0500000US01007, 0500000US01009, 0500000US01011, 0500000US01013, 0500000US01015, 0500...
$ GEOID    <fct> 01001, 01003, 01005, 01007, 01009, 01011, 01013, 01015, 01017, 01019, 01021, 01023, 01025, 01027, 01029, 01031, 01033, 01035, 01037,...
$ NAME     <fct> Autauga, Baldwin, Barbour, Bibb, Blount, Bullock, Butler, Calhoun, Chambers, Cherokee, Chilton, Choctaw, Clarke, Clay, Cleburne, Cof...
$ LSAD     <fct> 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, 06, ...
$ ALAND    <fct> 1539602123, 4117546676, 2292144655, 1612167481, 1670103911, 1613059160, 2012002530, 1569189622, 1545085607, 1433623321, 1794477075, ...
$ AWATER   <fct> 25706961, 1133055836, 50538698, 9602089, 15015423, 6054988, 2701198, 16627597, 16971701, 120308339, 20592144, 19144469, 36657891, 52...
$ geometry <MULTIPOLYGON [m]> MULTIPOLYGON (((1225889 -12..., MULTIPOLYGON (((1151155 -15..., MULTIPOLYGON (((1355398 -13..., MULTIPOLYGON (((1174172..
}
}
\examples{
library(Rnssp)
data("county_sf")

plot(county_sf$geometry)
}
\keyword{datasets}
