% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utilities.R
\name{create_apikey_profile}
\alias{create_apikey_profile}
\title{Create a profile with an API key}
\usage{
create_apikey_profile(
  api_key = askpass:::readline_silent("Enter/Paste an API key: "),
  key_name = "API-KEY"
)
}
\arguments{
\item{api_key}{a string for API key}

\item{key_name}{a string for an API Key name. Default is \code{API-KEY}}
}
\value{
An object of class \code{Apikey}
}
\description{
A wrapper to the \code{new} method defined in the \code{\link[Rnssp]{Apikey}} class.
}
\examples{
\dontrun{
# Create a profile with API key named "API-KEY"
myProfile <- create_apikey_profile()
myProfile

# Create a profile with an API key named "key"
myProfile2 <- create_apikey_profile(key_name = "key")
myProfile2
}

}
