% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utilities.R
\name{create_profile}
\alias{create_profile}
\title{A wrapper to the \code{new} method defined in the \code{\link[Rnssp]{Credentials}} class.}
\usage{
create_profile(
  username = askme("Please enter your username: "),
  password = askme()
)
}
\arguments{
\item{username}{a character of ESSENCE username}

\item{password}{a character of ESSENCE password}
}
\value{
an object of class NSSPCredentials and R6
}
\description{
A wrapper to the \code{new} method defined in the \code{\link[Rnssp]{Credentials}} class.
}
