% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utilities.R
\name{create_token_profile}
\alias{create_token_profile}
\title{A wrapper to the \code{new} method defined in the \code{\link[Rnssp]{Token}} class.}
\usage{
create_token_profile(
  token = askpass:::readline_silent("Enter/Paste a token: "),
  auth_type = "Bearer"
)
}
\arguments{
\item{token}{a string for token}

\item{auth_type}{type of HTTP authentication.
Should be \code{Bearer} or \code{Basic}. Default is \code{Bearer}}
}
\value{
An object of class \code{Token}
}
\description{
A wrapper to the \code{new} method defined in the \code{\link[Rnssp]{Token}} class.
}
\examples{
# Create a profile with "Bearer" authentication type
myProfile <- create_token_profile()
myProfile

# Create a profile with a "Basic" authentication type
myProfile2 <- create_token_profile(auth_type = "Basic")
myProfile2

}
