% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnssp_addins.R
\name{create_user_profile}
\alias{create_user_profile}
\title{User profile skeleton}
\usage{
create_user_profile()
}
\description{
Generate a user profile skeleton script into the console.
}
\keyword{internal}
