% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_switch.R
\name{detection_reg}
\alias{detection_reg}
\title{Fit linear regression model with time and day-of-week terms}
\usage{
detection_reg(x, algorithm = "regression", B, g)
}
\arguments{
\item{x}{Baselines of length B for each individual test date}

\item{algorithm}{Specifies detection algorithm. Defaults to "regression".}

\item{B}{Baseline parameter. The baseline length is the number of days to
which each linear model is fit. Defaults to 28 days.}

\item{g}{Guardband parameter. Number of days in buffer period to separate
the test date from the baseline.}
}
\value{
\itemize{
     \item If algorithm is "regression", numeric value of test statistic for
     one-sided Student's t-test.

     \item If algorithm is "switch", a character string specifying the fitted value,
      test statistic, and adjusted R-squared value to determine if the switch
      to EWMA is made.
}
}
\description{
Fit linear regression model with time and day-of-week terms
}
\keyword{internal}
