% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_farrington.R
\name{farrington_modified}
\alias{farrington_modified}
\title{Modified Farrington Algorithm}
\usage{
farrington_modified(df, t = date, y = count, B = 4, g = 27, w = 3, p = 10)
}
\arguments{
\item{df}{Nested dataframe}

\item{t}{A column containing date values}

\item{y}{A column containing time series counts}

\item{B}{Number of years to include in baseline (default is 4)}

\item{g}{Number of guardband weeks to separate the test date from the baseline (default is 27)}

\item{w}{Half the number of weeks included in reference window, before and after each reference date (default is 3)}

\item{p}{Number of seasonal periods for each year in baseline}
}
\value{
A tibble
}
\description{
Modified Farrington Algorithm
}
\keyword{internal}
