% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utilities.R
\name{get_api_data}
\alias{get_api_data}
\title{Get API data}
\usage{
get_api_data(url, fromCSV = FALSE, profile = myProfile, ...)
}
\arguments{
\item{url}{a character of API URL.}

\item{fromCSV}{a logical, defines whether data are returned in .csv format or .json format.}

\item{profile}{an object of class \code{\link[Rnssp]{Auth}}. Default is \code{myProfile}.}

\item{...}{further arguments and CSV parsing parameters to be passed to \code{\link[readr]{read_csv}} when \code{fromCSV = TRUE}.}
}
\value{
a dataframe (\code{fromCSV = TRUE}) or a list containing a dataframe and its metadata (\code{fromCSV = TRUE}).
}
\description{
A wrapper to the \code{get_api_data} method defined in the \code{\link[Rnssp]{Credentials}},
\code{\link[Rnssp]{Token}}, and \code{\link[Rnssp]{Apikey}} classes.
}
\examples{
\dontrun{
myProfile <- create_profile("", "")

json_url <- "http://httpbin.org/json"
api_data_json <- get_api_data(json_url)
head(api_data_json$slideshow$slides)

csv_url <- "http://httpbin.org/robots.txt"
api_data_csv <- get_api_data(csv_url, fromCSV = TRUE)
head(api_data_csv)
}
}
\seealso{
\code{\link[Rnssp]{get_api_response}} and \code{\link[Rnssp]{get_api_graph}}
}
