% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utilities.R
\name{get_api_graph}
\alias{get_api_graph}
\title{Get an API graph}
\usage{
get_api_graph(url, file_ext = ".png", profile = myProfile, ...)
}
\arguments{
\item{url}{a character of API URL.}

\item{file_ext}{a non-empty character vector giving the file extension.
Default is \code{.png}.}

\item{profile}{an object of class \code{\link[Rnssp]{Credentials}}.
Default is \code{myProfile}.}
}
\value{
An object of class \code{response}.
}
\description{
A wrapper to the \code{get_api_graph} method defined in the
\code{\link[Rnssp]{Credentials}} class.
}
\examples{
\dontrun{
myProfile <- create_profile("", "")
url <- "http://httpbin.org/image/png"

api_data_graph <- get_api_graph(url)

names(api_data_graph)
img <- png::readPNG(api_data_graph$graph)
grid::grid.raster(img)
}
}
\seealso{
\code{\link[Rnssp]{Credentials}}
}
