% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utilities.R
\name{get_essence_data}
\alias{get_essence_data}
\title{Get ESSENCE data}
\usage{
get_essence_data(
  url,
  start_date = NULL,
  end_date = NULL,
  profile = myProfile,
  ...
)
}
\arguments{
\item{url}{a character of ESSENCE API URL.}

\item{start_date}{a date object or a character string in date format (e.g. "2019-08-01")}

\item{end_date}{a date object or a character string in date format (e.g. "2020-08-01")}

\item{profile}{an object of class \code{\link[Rnssp]{Credentials}}. Default is \code{myProfile}.}

\item{...}{further arguments and CSV parsing parameters to be passed to \code{\link[readr]{read_csv}} when \code{fromCSV = TRUE}.}
}
\value{
a dataframe or a character string.
}
\description{
Get ESSENCE API data.
}
\examples{

}
\seealso{
\code{\link[Rnssp]{get_api_data}} and \code{\link[Rnssp]{get_api_tsgraph}}
}
