% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{list_apps}
\alias{list_apps}
\title{List Available Shinyapps}
\usage{
list_apps(as.table = FALSE)
}
\arguments{
\item{as.table}{a logical, if TRUE, a data frame is returned.
Otherwise, a vector is returned (Default is FALSE).}
}
\value{
A data frame or a vector
}
\description{
List available NSSP Shinyapps from the Rnssp-shiny-apps Github repository.
}
\examples{
list_apps()
list_apps(as.table = TRUE)
}
\seealso{
\href{https://github.com/CDCgov/Rnssp-shiny-apps}{https://github.com/CDCgov/Rnssp-rmd-templates}
}
