% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{list_templates}
\alias{list_templates}
\title{List Available Report Templates}
\usage{
list_templates(as.table = FALSE)
}
\arguments{
\item{as.table}{a logical, if TRUE, a data frame is returned.
Otherwise, a vector is returned (Default is FALSE).}
}
\value{
A data frame or a vector
}
\description{
List available NSSP report templates from the Rnssp-rmd-templates Github repository.
}
\examples{
list_templates()
list_templates(as.table = TRUE)
}
\seealso{
\code{\link{https://github.com/CDCgov/Rnssp-rmd-templates}}
}
