% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{NSSPContainer}
\alias{NSSPContainer}
\title{A \code{NSSPCredentials} Class to store a value in a container object}
\description{
A \code{NSSPCredentials} object stores a value
}
\details{
A \code{NSSPCredentials} object can be used to retrieve a value
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NSSPContainer-new}{\code{NSSPContainer$new()}}
\item \href{#method-NSSPContainer-clone}{\code{NSSPContainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPContainer-new"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPContainer-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new NSSPContainer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NSSPContainer$new(entry)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entry}}{an object or value to be stored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{NSSPCredentials} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPContainer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPContainer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NSSPContainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
