% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{remove_rmd_template}
\alias{remove_rmd_template}
\title{Utility to Remove an Existing RMarkdown Template}
\usage{
remove_rmd_template(
  template,
  pkg = "Rnssp",
  recursive = TRUE,
  force = TRUE,
  restart = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{template}{a character string with the name of the template to delete}

\item{pkg}{a character string with the name of a single package.
An error occurs if more than one package name is given.}

\item{recursive}{logical. Should directories be deleted recursively? (Default is TRUE)}

\item{force}{logical. Should permissions be changed (if possible) to allow the file
or directory to be removed? (Default is TRUE)}

\item{restart}{logical. Should R session be restarted after template
removal? (default is TRUE)}

\item{verbose}{logical. Should alert session restart message be printed?}
}
\value{
a character string
}
\description{
Remove an Existing Rnssp Rmarkdown template directory.
}
\examples{
\dontrun{
remove_rmd_template("text_mining") # Remove the Existing Rnssp 'text_mining' template
}
}
