% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnssp_gui.R
\name{remove_rmd_template_gui}
\alias{remove_rmd_template_gui}
\title{Remove Rnssp templates}
\usage{
remove_rmd_template_gui()
}
\description{
Wrapper around \code{\link[=remove_rmd_template]{remove_rmd_template()}}
to Remove Existing NSSP RMarkdown Templates.
}
\keyword{internal}
