% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnssp-package.R
\docType{package}
\name{Rnssp-package}
\alias{Rnssp-package}
\title{Rnssp: A Signature R package for the National Syndromic Surveillance Program (NSSP)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of tools, functions, and Rmarkdown templates that supports the National Syndromic and Surveillance Program Community of Practice.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cdcgov.github.io/Rnssp}
  \item \url{https://github.com/CDCgov/Rnssp}
}

}
\author{
\strong{Maintainer}: Gbedegnon Roseric Azondekon \email{gazondekon@cdc.gov} (\href{https://orcid.org/0000-0001-9155-4153}{ORCID}) [contributor]

Authors:
\itemize{
  \item Michael Sheppard \email{msheppard@cdc.gov} (\href{https://orcid.org/0000-0003-3660-575X}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Andrew Manigault \email{andrewmanigaultw@gmail.com} (\href{https://orcid.org/0000-0003-4197-1295}{ORCID}) [contributor]
}

}
\keyword{internal}
