% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnssp_stopwords.R
\docType{data}
\name{rnssp_stopwords}
\alias{rnssp_stopwords}
\title{Rnssp Stopwords}
\format{
A data frame with 835 rows and 2 variables:

\preformatted{
Rows: 835
Columns: 2
$ word <chr> "ABNORMAL", "ACCESS", "ACHE", "ACHES", "ACTIVE", "ACUTE", "ADMINISTERED", "ADVICE", "AFFECTED", "AFFORD", "AFTERWARDS", "AGO", "ALMOST", …
$ type <chr> "Medical Process Terms", "Medical Process Terms", "Medical Process Terms", "Medical Process Terms", "Medical Process Terms", "Medical Pro…
}
}
\usage{
rnssp_stopwords
}
\description{
A dataset containing:
\itemize{
    \item Medical Process Terms: Procedural terms used in medical setting (examples: administered, tested, treatment)
    \item Stop Words: Standard English stopwords (examples: some, then, would)
    \item Ignored Terms: Custom list of uninformative terms that frequently
          occur in syndromic surveillance data sources (examples: unspecified, encounter, midnight)
}
}
\examples{
library(Rnssp)
data("rnssp_stopwords")

glimpse(rnssp_stopwords)
}
\keyword{datasets}
