% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnssp_manual.R
\name{rnssp_templates_manual}
\alias{rnssp_templates_manual}
\title{Rnssp Templates Documentation}
\usage{
rnssp_templates_manual()
}
\description{
Browse the Rnssp R Markdown online documentation.
}
\keyword{internal}
