% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Rnssp_vignettes}
\alias{Rnssp_vignettes}
\title{View vignettes from the Rnssp package}
\usage{
Rnssp_vignettes(topic = NULL)
}
\arguments{
\item{topic}{a character string giving the (base) name of the vignette to view.
If omitted, the webpage with all vignettes from the Rnssp package is browsed}
}
\value{

}
\description{
Browse a specified Rnssp vignette, or browse the available ones
from the Rnssp online documentation.
}
\examples{
\dontrun{
Rnssp_vignettes()
Rnssp_vignettes(topic = "Rnssp_intro")
}
}
