% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run embedded shiny applications}
\usage{
run_app(app, launch_browser = TRUE)
}
\arguments{
\item{app}{A string representing the name of the application to run}

\item{launch_browser}{logical. Should the app be launch in the system's
default browser? (default is TRUE)}
}
\description{
Run embedded shiny applications
}
\examples{
run_app("rule_eval")
}
