% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnssp_addins.R
\name{run_app_gui}
\alias{run_app_gui}
\title{Run Rnssp Shinyapps}
\usage{
run_app_gui()
}
\description{
Wrapper around \code{\link[=run_app]{run_app()}} to run Rnssp Shinyapps from
the Rnssp Shinyapps Github repository.
}
\keyword{internal}
