% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_farrington.R
\name{seasonal_groups}
\alias{seasonal_groups}
\title{Return 10-level seasonal factor vector}
\usage{
seasonal_groups(B = 4, g = 27, w = 3, p = 10, base_length, base_weeks)
}
\arguments{
\item{B}{Number of years to include in baseline (default is 4)}

\item{g}{Number of guardband weeks to separate the test week from the
baseline (default is 27)}

\item{w}{Half the number of weeks included in reference window,
before and after each reference date (default is 3)}

\item{p}{Number of seasonal periods for each year in baseline (default is 10)}

\item{base_length}{Total number of weeks included in baseline}

\item{base_weeks}{Indices of baseline weeks}
}
\value{
A factor vector
}
\description{
Return 10-level seasonal factor vector
}
\keyword{internal}
