% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_serfling.R
\name{serfling_model}
\alias{serfling_model}
\title{Original Serfling method for weekly time series}
\usage{
serfling_model(df, t, y, baseline_end)
}
\arguments{
\item{df}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame}

\item{t}{Name of the column of type Date containing the dates}

\item{y}{Name of the column of type Numeric containing counts or
percentages}

\item{baseline_end}{Object of type Date defining the end of the
baseline/training period}
}
\value{
A data frame.
}
\description{
Serfling model helper function for monitoring weekly time series with
seasonality
}
\keyword{internal}
