% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated Time Series data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 626 rows and 5 columns.
}
\usage{
simulated_data
}
\description{
This dataset is a simulated time series for testing purposes.
}
\details{
\preformatted{
Rows: 626
Columns: 5
$ date  <date> 2017-01-01, 2017-01-08, 2017-01-15, 2017-01-22, 2017-01-29, 2017-02-0~
$ week  <dbl> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20,~
$ year  <dbl> 2017, 2017, 2017, 2017, 2017, 2017, 2017, 2017, 2017, 2017, 2017, 2017~
$ cases <dbl> 15, 12, 13, 11, 14, 16, 20, 15, 22, 19, 11, 14, 8, 6, 8, 12, 8, 14, 6,~
$ id    <chr> "Scenario #1", "Scenario #1", "Scenario #1", "Scenario #1", "Scenario ~
}
}
\examples{
library(Rnssp)
data("simulated_data")

head(simulated_data)
}
\keyword{datasets}
