% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_sf.R
\docType{data}
\name{state_sf}
\alias{state_sf}
\title{US States Shapefile}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 51 rows and 10 columns.
}
\usage{
state_sf
}
\description{
Useful \code{sf}` object shapefile for US States.
}
\details{
\preformatted{
Rows: 51
Columns: 10
$ STATEFP  <fct> 28, 37, 40, 51, 54, 22, 26, 25, 16, 12, 31, 53, 35, 46, 48, 06, 01, 13, 42, 29, 08, 49, 47, 56, 36, 20, 32, 17, 50, 30, 19, 45, 33, ...
$ STATENS  <fct> 01779790, 01027616, 01102857, 01779803, 01779805, 01629543, 01779789, 00606926, 01779783, 00294478, 01779792, 01779804, 00897535, 01...
$ AFFGEOID <fct> 0400000US28, 0400000US37, 0400000US40, 0400000US51, 0400000US54, 0400000US22, 0400000US26, 0400000US25, 0400000US16, 0400000US12, 04...
$ GEOID    <fct> 28, 37, 40, 51, 54, 22, 26, 25, 16, 12, 31, 53, 35, 46, 48, 06, 01, 13, 42, 29, 08, 49, 47, 56, 36, 20, 32, 17, 50, 30, 19, 45, 33, ...
$ STUSPS   <fct> MS, NC, OK, VA, WV, LA, MI, MA, ID, FL, NE, WA, NM, SD, TX, CA, AL, GA, PA, MO, CO, UT, TN, WY, NY, KS, NV, IL, VT, MT, IA, SC, NH, ...
$ NAME     <fct> Mississippi, North Carolina, Oklahoma, Virginia, West Virginia, Louisiana, Michigan, Massachusetts, Idaho, Florida, Nebraska, Washin...
$ LSAD     <fct> 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, 00, ...
$ ALAND    <fct> 121533519481, 125923656064, 177662925723, 102257717110, 62266474513, 111897594374, 146600952990, 20205125364, 214049787659, 13894913...
$ AWATER   <fct> 3926919758, 13466071395, 3374587997, 8528531774, 489028543, 23753621895, 103885855702, 7129925486, 2391722557, 31361101223, 13718291...
$ geometry <MULTIPOLYGON [m]> MULTIPOLYGON (((826504.7 -1..., MULTIPOLYGON (((1624281 -75..., MULTIPOLYGON (((-267188.2 -..., MULTIPOLYGON (((1578068...
}
}
\examples{
library(Rnssp)
data("state_sf")

plot(state_sf$geometry)
}
\keyword{datasets}
