% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add_rmd_template}
\alias{add_rmd_template}
\title{Prompt Utility to Add a New RMarkdown Template}
\usage{
add_rmd_template(pkg = "Rnssp")
}
\arguments{
\item{pkg}{a character string with the name of a single package. An error occurs if more than one package name is given.}
}
\value{
a character string
}
\description{
Prompt the user to select an Rmarkdown template .zip file.
}
\examples{
\dontrun{
add_rmd_template() # Add a new Rmd template to the 'Rnssp' package
add_rmd_template("rmarkdown") # Add a new Rmd template to the 'rmarkdown' package
}
}
