% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apikey.R
\name{Apikey}
\alias{Apikey}
\title{An \code{Apikey} Class Representing an Apikey object}
\description{
A \code{Apikey} object has an api_key string and a key name string.
}
\details{
An \code{Apikey} object can get API data via an API URL.
}
\examples{

## ------------------------------------------------
## Method `Apikey$get_api_response`
## ------------------------------------------------

\dontrun{
myProfile <- Apikey$new("abc1234567890")
url <- "https://httpbin.org/json"
api_response <- myProfile$get_api_response(url)
}

## ------------------------------------------------
## Method `Apikey$get_api_graph`
## ------------------------------------------------

\dontrun{
myProfile <- Apikey$new("abc1234567890")
url <- "https://httpbin.org/image/png"
api_data_graph <- myProfile$get_api_graph(url)
names(api_data_graph)
img <- png::readPNG(api_data_graph$graph)
grid::grid.raster(img)
}
}
\section{Super class}{
\code{Rnssp::NSSPAuth} -> \code{NSSPApikey}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NSSPApikey-new}{\code{Apikey$new()}}
\item \href{#method-NSSPApikey-get_api_response}{\code{Apikey$get_api_response()}}
\item \href{#method-NSSPApikey-get_api_graph}{\code{Apikey$get_api_graph()}}
\item \href{#method-NSSPApikey-clone}{\code{Apikey$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Rnssp" data-topic="NSSPAuth" data-id="get_api_data"><a href='../../Rnssp/html/NSSPAuth.html#method-NSSPAuth-get_api_data'><code>Rnssp::NSSPAuth$get_api_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPApikey-new"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPApikey-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new Credentials object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apikey$new(api_key, key_name = "API-KEY")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{a string for API key}

\item{\code{key_name}}{a string for an API Key name. Default is \code{API-KEY}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{Apikey}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPApikey-get_api_response"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPApikey-get_api_response}{}}}
\subsection{Method \code{get_api_response()}}{
Get API response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apikey$get_api_response(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{a character of API URL}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{response}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
myProfile <- Apikey$new("abc1234567890")
url <- "https://httpbin.org/json"
api_response <- myProfile$get_api_response(url)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPApikey-get_api_graph"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPApikey-get_api_graph}{}}}
\subsection{Method \code{get_api_graph()}}{
Get API graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apikey$get_api_graph(url, file_ext = ".png")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{a character of API URL}

\item{\code{file_ext}}{a non-empty character vector giving the file extension. Default is \code{.png}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing an api_response object and a path to a time series graph in .png format
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
myProfile <- Apikey$new("abc1234567890")
url <- "https://httpbin.org/image/png"
api_data_graph <- myProfile$get_api_graph(url)
names(api_data_graph)
img <- png::readPNG(api_data_graph$graph)
grid::grid.raster(img)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPApikey-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPApikey-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Apikey$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
