% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{askme}
\alias{askme}
\title{Password Prompt Utility}
\usage{
askme(prompt = "Please enter your password: ")
}
\arguments{
\item{prompt}{the string printed when prompting the user for input.}
}
\value{
a character string
}
\description{
Prompt the user for a password. This function is a wrapper for the `askpass::askpass()` function.
}
\examples{
password <- askme()
}
\seealso{
\code{\link[askpass]{askpass}}
}
