% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{Credentials}
\alias{Credentials}
\title{A \code{Credentials} Class Representing a Credentials object}
\description{
A \code{Credentials} object has a username and a password.
}
\details{
A \code{Credentials} object can get API data via an API URL.
}
\examples{

## ------------------------------------------------
## Method `Credentials$get_api_response`
## ------------------------------------------------

\dontrun{
myProfile <- Credentials$new(askme("Enter my username: "), askme())
url <- "https://httpbin.org/json"
api_response <- myProfile$get_api_response(url)
}

## ------------------------------------------------
## Method `Credentials$get_api_graph`
## ------------------------------------------------

\dontrun{
myProfile <- Credentials$new(askme("Enter my username: "), askme())
url <- "<url>"
api_data_graph <- myProfile$get_api_graph(url)
names(api_data_graph)
img <- png::readPNG(api_data_graph$graph)
grid::grid.raster(img)
}
}
\section{Super class}{
\code{Rnssp::NSSPAuth} -> \code{NSSPCredentials}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NSSPCredentials-new}{\code{Credentials$new()}}
\item \href{#method-NSSPCredentials-get_api_response}{\code{Credentials$get_api_response()}}
\item \href{#method-NSSPCredentials-get_api_graph}{\code{Credentials$get_api_graph()}}
\item \href{#method-NSSPCredentials-clone}{\code{Credentials$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Rnssp" data-topic="NSSPAuth" data-id="get_api_data"><a href='../../Rnssp/html/NSSPAuth.html#method-NSSPAuth-get_api_data'><code>Rnssp::NSSPAuth$get_api_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPCredentials-new"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPCredentials-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new Credentials object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$new(username = NULL, password = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{a string for username}

\item{\code{password}}{a string for password}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Credentials} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPCredentials-get_api_response"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPCredentials-get_api_response}{}}}
\subsection{Method \code{get_api_response()}}{
Get API response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$get_api_response(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{a character of API URL}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{response}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
myProfile <- Credentials$new(askme("Enter my username: "), askme())
url <- "https://httpbin.org/json"
api_response <- myProfile$get_api_response(url)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPCredentials-get_api_graph"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPCredentials-get_api_graph}{}}}
\subsection{Method \code{get_api_graph()}}{
Get API graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$get_api_graph(url, file_ext = ".png")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{a character of API URL}

\item{\code{file_ext}}{a non-empty character vector giving the file extension. Default is \code{.png}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing an api_response object and a path to a time series graph in .png format
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
myProfile <- Credentials$new(askme("Enter my username: "), askme())
url <- "<url>"
api_data_graph <- myProfile$get_api_graph(url)
names(api_data_graph)
img <- png::readPNG(api_data_graph$graph)
grid::grid.raster(img)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NSSPCredentials-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NSSPCredentials-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Credentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
