% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_reg.R
\name{detection_reg}
\alias{detection_reg}
\title{Fit linear regression model with time and day-of-week terms}
\usage{
detection_reg(x, algorithm = "regression", B, g)
}
\arguments{
\item{x}{A datatable object.}

\item{algorithm}{Specifies detection algorithm. Defaults to "regression".}

\item{B}{Baseline parameter. The baseline length is the number of days to
which each linear model is fit. Defaults to 28 days.}

\item{g}{Guardband parameter. Number of days in buffer period to separate
the test date from the baseline.}
}
\value{
- A numeric value of test statistic for one-sided Student's t-test when
      algorithm is "regression"
    - A character string specifying the fitted value, test statistic, and
      adjusted R-squared when algorithm is "switch".
}
\description{
Fit linear regression model with time and day-of-week terms
}
