% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_ewma.R
\name{ewma_loop}
\alias{ewma_loop}
\title{Loop for EWMA}
\usage{
ewma_loop(y, mu, sigma1, sigma2, B, g, w, alert_thresh)
}
\arguments{
\item{y}{Numeric vector of counts or percentages}

\item{mu}{Numeric vector of baseline averages}

\item{sigma1}{Adjusted standard deviation for baselines and
first weighting for sensitivity to gradual events}

\item{sigma2}{Adjusted standard deviation for baselines and
second weighting for sensitivity to sudden events}

\item{w}{Vector of length 2 with smoothing coefficients for
gradual and sudden events}

\item{alert_thresh}{Alert threshold value}
}
\value{
List of output p-values, test statistics, and EWMAs
}
\description{
Loop for EWMA and adjustment of outlying smoothed values
}
\keyword{internal}
