% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run embedded shiny applications}
\usage{
run_app(app)
}
\arguments{
\item{app}{A string representing the name of the application to run}
}
\value{

}
\description{
Run embedded shiny applications
}
